fix: remove onClick handler for non-actionable elements (#18113)
* fix: remove onClick handler for non-actionable elements

* chore: add changelog

* fix: add Unit Test
